requirements.txt
asgiref==3.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
drf-spectacular==0.21.1
importlib-metadata==4.10.0
importlib-resources==5.4.0
inflection==0.5.1
jsonschema==4.4.0
Markdown==3.3.6
Pillow==9.0.0
pyrsistent==0.18.1
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
zipp==3.7.0
